fix(codeql): clear the 11 open alerts (10 high) — casts + a workflow permissions block (#2340)
CodeQL has been carrying **11 open alerts**, 10 of them HIGH, and nobody has
triaged them. Clearing noise matters here because a real finding would be lost in
it: the same rule on `engine/npu/src/dequant_q4nx.cpp` is already marked *fixed*,
so fixing rather than dismissing is the house style.

**10 × `cpp/integer-multiplication-cast-to-long` in
`engine/npu/tools/lfm2_cpu_runner.cpp`** (added today in #2316, so the alerts are
hours old): 7 sites where an `int * int` product is widened to the container's
`size_type` — 4 in the `qd/kd/vd/attn` vector sizing (line 284) and 3 in the
per-head indexing (`&qd[hh * HD]`, `&kd[hh * HD]`, `&attn[hh * HD]`). Each now
casts the first operand to `size_t`, so the multiplication happens in the wide
type. Behaviour is unchanged at realistic dimensions (`NH*HD` = 4096; the largest,
`pos*NKV*HD`, is ~8.4M at 8k context) — this is hardening, not a live bug, and I
am not claiming otherwise.

**1 × `actions/missing-workflow-permissions` in `.github/workflows/deploy.yml`** —
the `indexnow` job declared no `permissions:` and so inherited the repository
default for its token, while all it does is read the repo and POST URLs. It now
declares `contents: read`. The other two jobs already declared theirs and are
untouched (`deploy`: contents read + deployments write; `deploy-github-pages`:
contents read, pages write, id-token write).

Verification:

- `g++ -fsyntax-only -std=c++23 … lfm2_cpu_runner.cpp` → exit 0
- the workflow parses and the three jobs' effective permissions are as listed above

Process note: GitNexus `impact main --direction upstream` was ambiguous (many
`main`s), which the rules say is not an all-clear, so the blast radius was
confirmed by text search instead — nothing in the tree references
`lfm2_cpu_runner` (no build script, no CMake, no docs), i.e. it is a standalone
tool with no callers.

+++ engine/npu/tools/lfm2_cpu_runner.cpp
@@ -281,7 +281,11 @@ int main(int argc, char** argv) {
     fprintf(stderr, "lm_head: %zu floats (N=%d K=%d)\n", lm_head.v.size(), NV, H);
 
     std::vector<float> h(H), x(H), res(H), tmp(H), tmp2(H);
-    std::vector<float> qd(NH * HD), kd(NKV * HD), vd(NKV * HD), attn(NH * HD);
+    // (size_t) on the first operand: these products are widened to the vector's
+    // size_type but the multiply itself happens in int — CodeQL
+    // cpp/integer-multiplication-cast-to-long, 4 of the sites.
+    std::vector<float> qd((size_t)NH * HD), kd((size_t)NKV * HD),
+        vd((size_t)NKV * HD), attn((size_t)NH * HD);
     int last_tok = -1;
 
     std::vector<int> gen;
@@ -362,10 +366,10 @@ int main(int argc, char** argv) {
                 mul(wq, x.data(), qd.data());
                 mul(wk, x.data(), kd.data());
                 mul(wv, x.data(), vd.data());
-                for (int hh = 0; hh < NH; hh++) rmsnorm(&qd[hh * HD], qn.data(), HD, eps);
-                for (int hh = 0; hh < NKV; hh++) rmsnorm(&kd[hh * HD], kn.data(), HD, eps);
-                for (int hh = 0; hh < NH; hh++) rope(&qd[hh * HD], HD, pos, ROPE_THETA);
-                for (int hh = 0; hh < NKV; hh++) rope(&kd[hh * HD], HD, pos, ROPE_THETA);
+                for (int hh = 0; hh < NH; hh++) rmsnorm(&qd[(size_t)hh * HD], qn.data(), HD, eps);
+                for (int hh = 0; hh < NKV; hh++) rmsnorm(&kd[(size_t)hh * HD], kn.data(), HD, eps);
+                for (int hh = 0; hh < NH; hh++) rope(&qd[(size_t)hh * HD], HD, pos, ROPE_THETA);
+                for (int hh = 0; hh < NKV; hh++) rope(&kd[(size_t)hh * HD], HD, pos, ROPE_THETA);
                 memcpy(&kv_k[l][(size_t)pos * NKV * HD], kd.data(), kv_k[l].size() ? (size_t)NKV * HD * 4 : 0);
                 memcpy(&kv_v[l][(size_t)pos * NKV * HD], vd.data(), kv_v[l].size() ? (size_t)NKV * HD * 4 : 0);
 
@@ -373,8 +377,8 @@ int main(int argc, char** argv) {
                 #pragma omp parallel for schedule(static)
                 for (int hh = 0; hh < NH; hh++) {
                     int kvh = hh / GQA;
-                    const float* q = &qd[hh * HD];
-                    float* out = &attn[hh * HD];
+                    const float* q = &qd[(size_t)hh * HD];
+                    float* out = &attn[(size_t)hh * HD];
                     std::vector<float> sc(pos + 1);
                     float mx = -1e30f;
                     for (int p2 = 0; p2 <= pos; p2++) {
